requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
comm==0.1.4
Cython==3.0.3
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
